+++ CHANGELOG.md
@@ -5,10 +5,21 @@
   * Digraph.new now yields self if a block is given. This allows building a 
     graph more smoothly:
     
-        Digraph.new{|d|
+        graph = Digraph.new{|d|
           d.add_n_vertices(5)
           ...
         }
+        
+  * An integer is now automatically recognized as a selection predicate. This 
+    means that the following will work:
+    
+        # [... previous example continued ...]
+        graph.vertices(2) # => returns the 2-th vertex (0, 1, **2**)
+        
+        # connection can be made simply as
+        graph.connect(1, 2)
+
+  * Added Digraph#ith_vertex and Digraph#ith_edge 
 
 * Internals & Devel